Spawner, controller, and state scripts
Here Is an implementation of one of my AI scripts. 

the AI_state script creates the base class all AI in the game were built off of, implementing the state machine, all possible states, and virtual functions for all needed functions.

The AI_controller script is poorly named but was named so because in the time given i was only able to implement one ai a basic crew member and test them against provided input from an outside player object for their reactablitiy and normal function outside of player interaction. 

the spawner script was created to create a pool of objects with randomly generated tasks for them to accomplish before returning and despawning.